+++ datanode/gateway/graphql/update_asset_resolver.go
@@ -24,14 +24,6 @@ func (r *updateAssetResolver) Source(ctx context.Context, obj *types.UpdateAsset
 	return UpdateAssetSourceFromProto(obj.Changes)
 }
 
-func (r updateAssetResolver) Name(ctx context.Context, obj *types.UpdateAsset) (string, error) {
-	return obj.Changes.Name, nil
-}
-
-func (r updateAssetResolver) Symbol(ctx context.Context, obj *types.UpdateAsset) (string, error) {
-	return obj.Changes.Symbol, nil
-}
-
 func (r *updateAssetResolver) Quantum(ctx context.Context, obj *types.UpdateAsset) (string, error) {
 	return obj.Changes.Quantum, nil
 }